+++ src/types.ts
@@ -1,5 +1,5 @@
 export enum TaskType {
-  CODE_STRUCTURE = 'analyze_code_structure',
+  ANALYZE_SINGLE_FILE = 'analyze_single_file',
   GENERATE_TESTS = 'generate_unit_tests',
   DOCUMENT_FUNCTION = 'generate_documentation',
   SUGGEST_REFACTOR = 'suggest_refactoring',